Remove ProtocolEncoder abstraction
