+++ packages/client/src/runtime/RequestHandler.ts
@@ -8,6 +8,7 @@ import {
   EventEmitter,
   Fetch,
   InteractiveTransactionOptions,
+  JsonQuery,
   TransactionOptions,
 } from '../runtime/core/engines'
 import {
@@ -23,14 +24,17 @@ import { createApplyBatchExtensionsFunction } from './core/extensions/applyQuery
 import { applyResultExtensions } from './core/extensions/applyResultExtensions'
 import { MergedExtensionsList } from './core/extensions/MergedExtensionsList'
 import { visitQueryResult } from './core/extensions/visitQueryResult'
+import { deserializeJsonResponse } from './core/jsonProtocol/deserializeJsonResponse'
+import { getBatchId } from './core/jsonProtocol/getBatchId'
+import { isWrite } from './core/jsonProtocol/isWrite'
 import { dmmfToJSModelName } from './core/model/utils/dmmfToJSModelName'
-import { ProtocolEncoder, ProtocolMessage } from './core/protocol/common'
 import { PrismaPromiseInteractiveTransaction, PrismaPromiseTransaction } from './core/request/PrismaPromise'
 import { Action, JsArgs } from './core/types/JsApi'
 import { DataLoader } from './DataLoader'
 import type { Client, Unpacker } from './getPrismaClient'
 import { CallSite } from './utils/CallSite'
 import { createErrorMessageWithContext } from './utils/createErrorMessageWithContext'
+import { deepGet } from './utils/deep-set'
 import { NotFoundError, RejectOnNotFound, throwIfNotFound } from './utils/rejectOnNotFound'
 
 const debug = Debug('prisma:client:request_handler')
@@ -38,8 +42,7 @@ const debug = Debug('prisma:client:request_handler')
 export type RequestParams = {
   modelName?: string
   action: Action
-  protocolMessage: ProtocolMessage
-  protocolEncoder: ProtocolEncoder
+  protocolQuery: JsonQuery
   dataPath: string[]
   clientMethod: string
   callsite?: CallSite
@@ -80,14 +83,14 @@ export class RequestHandler {
 
     this.dataloader = new DataLoader({
       batchLoader: createApplyBatchExtensionsFunction(async ({ requests, customDataProxyFetch }) => {
-        const { transaction, protocolEncoder, otelParentCtx } = requests[0]
-        const queries = protocolEncoder.createBatch(requests.map((r) => r.protocolMessage))
+        const { transaction, otelParentCtx } = requests[0]
+        const queries = requests.map((r) => r.protocolQuery)
         const traceparent = this.client._tracingHelper.getTraceParent(otelParentCtx)
 
         // TODO: pass the child information to QE for it to issue links to queries
         // const links = requests.map((r) => trace.getSpanContext(r.otelChildCtx!))
 
-        const containsWrite = requests.some((r) => r.protocolMessage.isWrite())
+        const containsWrite = requests.some((r) => isWrite(r.protocolQuery.action))
 
         const results = await this.client._engine.requestBatch(queries, {
           traceparent,
@@ -113,10 +116,10 @@ export class RequestHandler {
         const interactiveTransaction =
           request.transaction?.kind === 'itx' ? getItxTransactionOptions(request.transaction) : undefined
 
-        const response = await this.client._engine.request(request.protocolMessage.toEngineQuery(), {
+        const response = await this.client._engine.request(request.protocolQuery, {
           traceparent: this.client._tracingHelper.getTraceParent(),
           interactiveTransaction,
-          isWrite: request.protocolMessage.isWrite(),
+          isWrite: isWrite(request.protocolQuery.action),
           customDataProxyFetch: request.customDataProxyFetch,
         })
         return this.mapQueryEngineResult(request, response)
@@ -127,7 +130,7 @@ export class RequestHandler {
           return `transaction-${request.transaction.id}`
         }
 
-        return request.protocolMessage.getBatchId()
+        return getBatchId(request.protocolQuery)
       },
 
       batchOrder(requestA, requestB) {
@@ -151,7 +154,7 @@ export class RequestHandler {
   }
 
   mapQueryEngineResult(
-    { protocolMessage, dataPath, unpacker, modelName, args, extensions }: RequestParams,
+    { dataPath, unpacker, modelName, args, extensions }: RequestParams,
     response: QueryEngineResult<any>,
   ) {
     const data = response?.data
@@ -160,7 +163,7 @@ export class RequestHandler {
     /**
      * Unpack
      */
-    let result = this.unpack(protocolMessage, data, dataPath, unpacker)
+    let result = this.unpack(data, dataPath, unpacker)
     if (modelName) {
       result = this.applyResultExtensions({ result, modelName, args, extensions })
     }
@@ -256,7 +259,7 @@ export class RequestHandler {
     return message
   }
 
-  unpack(message: ProtocolMessage, data: unknown, dataPath: string[], unpacker?: Unpacker) {
+  unpack(data: unknown, dataPath: string[], unpacker?: Unpacker) {
     if (!data) {
       return data
     }
@@ -264,7 +267,13 @@ export class RequestHandler {
       data = data['data']
     }
 
-    const deserializeResponse = message.deserializeResponse(data, dataPath)
+    if (!data) {
+      return data
+    }
+    const response = Object.values(data)[0]
+    const pathForGet = dataPath.filter((key) => key !== 'select' && key !== 'include')
+    const deserializeResponse = deserializeJsonResponse(deepGet(response, pathForGet))
+
     return unpacker ? unpacker(deserializeResponse) : deserializeResponse
   }
