+++ packages/client/src/runtime/core/jsonProtocol/isWrite.ts
@@ -1,0 +1,27 @@
+import { JsonQueryAction } from '../engines'
+
+const writeMap: Record<JsonQueryAction, boolean> = {
+  aggregate: false,
+  aggregateRaw: false,
+  createMany: true,
+  createOne: true,
+  deleteMany: true,
+  deleteOne: true,
+  executeRaw: true,
+  findFirst: false,
+  findFirstOrThrow: false,
+  findMany: false,
+  findRaw: false,
+  findUnique: false,
+  findUniqueOrThrow: false,
+  groupBy: false,
+  queryRaw: false,
+  runCommandRaw: true,
+  updateMany: true,
+  updateOne: true,
+  upsertOne: true,
+}
+
+export function isWrite(action: JsonQueryAction): boolean {
+  return writeMap[action]
+}
