--- packages/client/src/runtime/core/protocol/json/index.ts
@@ -1,1 +1,0 @@
-export { JsonProtocolEncoder, JsonProtocolMessage } from './JsonProtocolEncoder'
